gcp_tutorial/requirements.txt
pandas
seaborn
google-cloud-bigquery
pyarrow
pandas-gbq
jupyterlab
streamlit